add xml config from rails-infrastructure-canary job
These were downloaded from these two urls:
http://jenkins.uscm.org/job/rails-infrastructure-canary/config.xml
http://jenkins.uscm.org/job/rails-infrastructure-canary/promotion/process/Deploy%20to%20Production/config.xml

The rails-infrastructure-canary job was manually set up in the Jenkins UI. It'll be the basis for developing a jerkins-job-builder yaml script for the various jobs in jenkins.
These scripts will eventually be removed from this repo, but until then they'll provide the "goal" that the yaml script needs to create.