only auto-build master and staging branches
This commit uses a regular expression to match only 'origin/master' and 'origin/staging'.
(The git plugin allows configuring two different branches,
but this doesn't play well with using a parameter to drive which branch to build.)

We've decided that there are enough drawbacks from running builds on every feature branch commit, (like jenkins job queuing)
and not enough benefits.