add a production release job
This fairly involved commit removes things that only work for snapshot deployments, and instead adds a job to perform a 'release'.
The release job calls a devops script which
 1. pulls down the commerce-engine and extensions repositories, and then
 2. reversions the the pom.xml files to the release number.
The job then builds and deploys both projects and creates git tags.

This new job takes the function of the give-ep-{commerce-engine,extensions,deployment-package}-{build,deploy}-master jobs, which are removed in this commit.
The give-ep-{commerce-engine,extensions}-test-master jobs will be kept, but they will be triggered from merges to master instead of the 'build' jobs.

The docker images and deployment jobs are fed the release number in order to grab the released deployment package from the nexus repository.

To deploy things to production, we will:
1. manually do a fast-forward merge from develop -> master for one or multiple of the extensions, devops, and commerce-engine repos
2. wait for tests to finish successfully (optionally?)
3. manually run the release job
4. automatically deploy the resulting docker image to UAT, to do any final testing there -- this is not implemented yet
5. manually run the deploy-prod job