ensure that prod EP promotion propagates RELEASE_NUMBER
For Staging, GIT_COMMIT/BUILD_NUMBER are what is needed by the deploy job, so the existing config works for staging.
But for Prod, the deploy job needs RELEASE_NUMBER instead of BUILD_NUMBER.

This commit adds RELEASE_NUMBER if it was passed in to the promotion, which in production it is. In stage, it's not, so this doesn't affect the staging promotion.

(It turns out that BUILD_NUMBER is also sent for the prod promotion, but that is fine; it is just ignored.)