go back to `predefined-parameters`
It turns out that the behavior is different when expressions (e.g. ${BUILD_NUMBER}) are involved.
With `current-parameters`, the _expression_ is passed to the next build. In our case, the next job's build number gets used, and not this job's.
With `predefined-parameters`, using syntax like this commit does, the evaluated _result_ of the expression is passed to the next build. In our case, this job's build number is passed on.