derive ert-api ENVIRONMENT from GIT_BRANCH
The rails apps do this via ecs_config's deploy.rb, but ert-api doesn't use that for image building/pushing.
The logic added here is derived from deploy.sh and deploy.rb.

This also adds some readability improvements.

Note: I couldn't find a simple way to determine and set ENVIRONMENT before the maven build. I can run a shell script, but it can't change the jenkins environment.
I could run a groovy snippet, but jenkins-job-builder doesn't support the config for that. This commit solves the problem by delaying the ENVIRONMENT determination until the post-build action, which is the only place that actually needs it.