add job for scale-of-belief
Note: I had originally planned to add `jenkins-jobs`
as a job library instead of a global library,
since it would save me the effort of setting up a global library with ansible.
It turned out that the groovy sandbox that is used for job libraries
is super restrictive, and finding and whitelisting methods is a painful process.
So that motivated me to switch using a global library,
even if it meant a little extra work in cru-ansible.