Add production deployment confirmation logic
Note 1: I had originally had Deployments as a class, instead of a 'script',
but I had to switch so that it could easily access Jenkins' 'echo' step.

Note 2: I had originally swapped the `stage('ConfirmDeployment')` a
and `if (deployments.afterHoursConfirmationRequired())) {` lines,
thinking it'd be helpful to explicitly remove the stage when it wasn't needed.

However, the Jenkins UI didn't handle this as nicely as I had hoped.
Instead of the build history page showing all the builds,
with some showing that stage and others not showing it,
it dropped the jobs that didn't have the same set of stages.
That is, it is determined to show a grid,
and it'll just drop historical builds that don't fit the grid.