Introduce deploy-{env}-job params to reduce copy/paste
This reworks and renames the docker-airflow-production job such that it
is structured more like deploy-ecs,
and introduces a parameter so that app-jobs can use a different
deployment job than deploy-ecs,
and configures docker-airflow to use the docker-airflow-deploy job
(renamed from docker-airflow-production.)

I renamed it since it's not particularly tied to production anymore,
except by the comments,
and also so that we can test it in parallel with the old job.

There is now a new form of copy/paste: the parameter list.
Might eventually try to deal with this via yaml includes,
but I'm not sure how hard that is.
In any case, this is less copy/paste than what was here before.