Have wsapi use the generic build flow
The current java build setup requires Java 7 or 8,
and requires that maven initiate the docker build.
Dmitry and I concluded that using maven to wrap docker isn't useful,
and we also decided to use Java 11 with wsapi.

So I think it's better to have wsapi provide a build.sh
that runs maven and the docker image build,
ignoring Jenkin's built-in support for maven projects.

I'm adding a 'generic-template' job here,
which will do exactly what the php-template job does:
execute build.sh or complain if it doesn't exist.
Probably I'll modify the php jobs to use this template
and remove php-template.